Move SavedPlanBookmark to dedicated `internal/cloud/cloudplan` package
Having this sitting loose in `cloud` proved problematic because of circular
import dependencies. Putting it in a sub-package under cloud frees us up to
reference the type from places like `internal/backend`!
