+++ internal/cloud/cloudplan/saved_plan.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package cloud
+package cloudplan
 
 import (
 	"encoding/json"
