+++ internal/cloud/cloudplan/testdata/plan-bookmark/bookmark.json
